Corrections on SynchronizeFacebookShop flag for batch calls.

+++ src/Modules/Laser.Orchard.NwazetIntegration/Handlers/FacebookShopProductPartHandler.cs
@@ -39,13 +39,15 @@ namespace Laser.Orchard.NwazetIntegration.Handlers {
         public Localizer T { get; set; }
 
         protected void AfterPublish(PublishContentContext context) {
-            if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
-                var result = _facebookShopService.SyncProduct(context.ContentItem);
+            if (context.ContentItem.As<FacebookShopProductPart>().SynchronizeFacebookShop) {
+                if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
+                    var result = _facebookShopService.SyncProduct(context.ContentItem);
 
-                if (result == null || result.Requests.Count == 0) {
-                    _notifier.Warning(T("Invalid Facebook api response. Product is not synchronised on Facebook catalog."));
-                } else if (!result.Requests[0].Valid) {
-                    _notifier.Warning(result.Requests[0].Message);
+                    if (result == null || result.Requests.Count == 0) {
+                        _notifier.Warning(T("Invalid Facebook api response. Product is not synchronised on Facebook catalog."));
+                    } else if (!result.Requests[0].Valid) {
+                        _notifier.Warning(result.Requests[0].Message);
+                    }
                 }
             }
         }
@@ -59,13 +61,15 @@ namespace Laser.Orchard.NwazetIntegration.Handlers {
         }
 
         private void RemoveProduct(ContentItem product) {
-            if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
-                var result = _facebookShopService.RemoveProduct(product);
+            if (product.As<FacebookShopProductPart>().SynchronizeFacebookShop) {
+                if (_authorizer.Authorize(FacebookShopSynchronizationPermission.FacebookShopSynchronization)) {
+                    var result = _facebookShopService.RemoveProduct(product);
 
-                if (result == null || result.Requests.Count == 0) {
-                    _notifier.Warning(T("Invalid Facebook api response. Product has not been removed from Facebook catalog."));
-                } else if (!result.Requests[0].Valid) {
-                    _notifier.Warning(result.Requests[0].Message);
+                    if (result == null || result.Requests.Count == 0) {
+                        _notifier.Warning(T("Invalid Facebook api response. Product has not been removed from Facebook catalog."));
+                    } else if (!result.Requests[0].Valid) {
+                        _notifier.Warning(result.Requests[0].Message);
+                    }
                 }
             }
         }
